refactor(shared)(api): move api error response into shared models package
